Made the app ugly by removing the material3 dependency and using the base theme.
TODO:
- Make my own theme that should hopefully make the app pretty :P
