+++ app/build.gradle.kts
@@ -43,9 +43,7 @@ android {
 dependencies {
 
     implementation(libs.appcompat)
-    implementation(libs.material)
     implementation(libs.activity)
-    implementation(libs.constraintlayout)
     implementation(libs.ui.android)
     testImplementation(libs.junit)
     androidTestImplementation(libs.ext.junit)
