+++ app/src/main/AndroidManifest.xml
@@ -21,7 +21,7 @@
         </provider>
 
         <activity
-            android:name=".Rework"
+            android:name=".MainActivity"
             android:exported="true" >
             <!--
             android:launchMode="singleTask"
