+++ app/src/main/res/drawable/textedit_background.xml
@@ -5,7 +5,7 @@
         <!-- pressed -->
         <shape android:shape="rectangle">
             <!--    <solid android:color="?attr/colorSurface" />-->
-            <stroke android:width="2dp" android:color="@color/design_default_color_primary_variant" />
+            <stroke android:width="2dp" android:color="?attr/colorPrimary" />
             <corners android:radius="8dp" />
         </shape>
     </item>
@@ -13,7 +13,7 @@
         <!-- focused -->
         <shape android:shape="rectangle">
             <!--    <solid android:color="?attr/colorSurface" />-->
-            <stroke android:width="2dp" android:color="@color/design_default_color_primary" />
+            <stroke android:width="2dp" android:color="?attr/colorAccent" />
             <corners android:radius="8dp" />
         </shape>
     </item>
@@ -21,7 +21,7 @@
         <!-- default -->
         <shape android:shape="rectangle">
             <!--    <solid android:color="?attr/colorSurface" />-->
-            <stroke android:width="1dp" android:color="?attr/cardBackgroundColor" />
+            <stroke android:width="1dp" android:color="?attr/colorControlHighlight" />
             <corners android:radius="8dp" />
         </shape>
     </item>
