+++ app/src/main/res/drawable/tool_tip_arrow.xml
@@ -4,7 +4,7 @@
     <item
         android:top="8dp" >
         <shape android:shape="rectangle">
-            <solid android:color="?attr/colorOnSecondary" />
+            <solid android:color="?attr/colorAccent" />
             <corners android:radius="8dp" />
             <padding android:top="8dp" />
         </shape>
@@ -20,7 +20,7 @@
             android:pivotY="82%"
             android:toDegrees="45">
             <shape android:shape="rectangle">
-                <solid android:color="?attr/colorOnSecondary" />
+                <solid android:color="?attr/colorAccent" />
 
             <!--<stroke-->
             <!--    android:width="2dp"-->
