+++ app/src/main/res/layout/activity_main.xml
@@ -1,48 +1,37 @@
 <?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MainActivity" >
+    android:orientation="vertical"
+    tools:context=".MainActivity">
+
 
     <ImageView
         android:id="@+id/imageViewQRCode"
         android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
+        android:layout_height="0dp"
+        android:layout_weight="0.5"
+        android:layout_gravity="center"
         android:contentDescription="@string/qr_code_description"
         app:autoSizeTextType="uniform"
         app:contentDescription="@string/qr_code_description"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintVertical_bias="0.3"
         app:srcCompat="@drawable/ic_launcher_foreground" />
 
-    <com.google.android.material.textfield.TextInputLayout
-        android:id="@+id/qr_subtitle_hint"
-        android:layout_width="0dp"
+    <EditText
+        android:id="@+id/qr_subtitle"
+        android:layout_width="match_parent"
         android:layout_height="0dp"
+        android:layout_weight="1"
         android:layout_margin="8dp"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/imageViewQRCode"
-        app:layout_constraintWidth_max="488dp" >
-        <!--488dp for material guidelines on width of text field-->
+        android:background="@drawable/textedit_background"
+        android:gravity="start"
+        android:hint="@string/qr_instructions"
+        android:inputType="text|textAutoComplete|textAutoCorrect|textEnableTextConversionSuggestions|textImeMultiLine|textMultiLine"
+        android:padding="8dp"
+        android:textAlignment="viewStart"
+        app:autoSizeTextType="uniform" />
 
-        <com.google.android.material.textfield.TextInputEditText
-            android:id="@+id/qr_subtitle"
-            android:layout_width="match_parent"
-            android:layout_height="match_parent"
-            android:gravity="start"
-            android:hint="@string/qr_instructions"
-            android:textAlignment="viewStart"
-            app:autoSizeTextType="uniform" />
-    </com.google.android.material.textfield.TextInputLayout>
-
-
-
-</androidx.constraintlayout.widget.ConstraintLayout>
+</LinearLayout>
