--- app/src/main/res/layout/activity_main2.xml
@@ -1,40 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/main"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".MainActivity" >
-
-    <ImageView
-        android:id="@+id/imageViewQRCode"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:contentDescription="@string/qr_code_description"
-        app:autoSizeTextType="uniform"
-        app:contentDescription="@string/qr_code_description"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintVertical_bias="0.3"
-        app:srcCompat="@drawable/ic_launcher_foreground" />
-
-
-    <EditText
-        android:id="@+id/qr_subtitle"
-        android:layout_width="match_parent"
-        android:layout_height="0dp"
-        android:hint="@string/qr_instructions"
-        android:inputType="text|textAutoComplete|textAutoCorrect|textEnableTextConversionSuggestions|textImeMultiLine|textMultiLine"
-        android:textAlignment="viewStart"
-        android:padding="8dp"
-        android:layout_margin="8dp"
-        android:background="@drawable/textedit_background"
-        android:gravity="start"
-        app:autoSizeTextType="uniform"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/imageViewQRCode" />
-
-</androidx.constraintlayout.widget.ConstraintLayout>
